Move use_source_factory and use_source_factory_metadata_index to source_factories
