+++ firecrown/likelihood/two_point.py
@@ -25,13 +25,15 @@ from firecrown.generators.two_point import (
     apply_theta_min_max,
 )
 from firecrown.likelihood.source import Source, Tracer
+from firecrown.likelihood.source_factories import (
+    use_source_factory,
+    use_source_factory_metadata_index,
+)
 from firecrown.likelihood.weak_lensing import (
     WeakLensingFactory,
-    WeakLensing,
 )
 from firecrown.likelihood.number_counts import (
     NumberCountsFactory,
-    NumberCounts,
 )
 from firecrown.likelihood.statistic import (
     DataVector,
@@ -39,9 +41,6 @@ from firecrown.likelihood.statistic import (
     TheoryVector,
 )
 from firecrown.metadata_types import (
-    Galaxies,
-    InferredGalaxyZDist,
-    Measurement,
     TRACER_NAMES_TOTAL,
     TracerNames,
     TwoPointHarmonic,
@@ -72,62 +71,6 @@ SACC_DATA_TYPE_TO_CCL_KIND = {
     "cmbGalaxy_convergenceDensity_xi": "NN",
     "cmbGalaxy_convergenceShear_xi_t": "NG",
 }
-
-
-def use_source_factory(
-    inferred_galaxy_zdist: InferredGalaxyZDist,
-    measurement: Measurement,
-    wl_factory: WeakLensingFactory | None = None,
-    nc_factory: NumberCountsFactory | None = None,
-) -> WeakLensing | NumberCounts:
-    """Apply the factory to the inferred galaxy redshift distribution."""
-    source: WeakLensing | NumberCounts
-    if measurement not in inferred_galaxy_zdist.measurements:
-        raise ValueError(
-            f"Measurement {measurement} not found in inferred galaxy redshift "
-            f"distribution {inferred_galaxy_zdist.bin_name}!"
-        )
-
-    match measurement:
-        case Galaxies.COUNTS:
-            assert nc_factory is not None
-            source = nc_factory.create(inferred_galaxy_zdist)
-        case (
-            Galaxies.SHEAR_E
-            | Galaxies.SHEAR_T
-            | Galaxies.SHEAR_MINUS
-            | Galaxies.SHEAR_PLUS
-        ):
-            assert wl_factory is not None
-            source = wl_factory.create(inferred_galaxy_zdist)
-        case _:
-            raise ValueError(f"Measurement {measurement} not supported!")
-    return source
-
-
-def use_source_factory_metadata_index(
-    sacc_tracer: str,
-    measurement: Measurement,
-    wl_factory: WeakLensingFactory | None = None,
-    nc_factory: NumberCountsFactory | None = None,
-) -> WeakLensing | NumberCounts:
-    """Apply the factory to create a source from metadata only."""
-    source: WeakLensing | NumberCounts
-    match measurement:
-        case Galaxies.COUNTS:
-            assert nc_factory is not None
-            source = nc_factory.create_from_metadata_only(sacc_tracer)
-        case (
-            Galaxies.SHEAR_E
-            | Galaxies.SHEAR_T
-            | Galaxies.SHEAR_MINUS
-            | Galaxies.SHEAR_PLUS
-        ):
-            assert wl_factory is not None
-            source = wl_factory.create_from_metadata_only(sacc_tracer)
-        case _:
-            raise ValueError(f"Measurement {measurement} not supported!")
-    return source
 
 
 class TwoPoint(Statistic):
